- Call Quantitation service 4.0  + config version 3.0 - Change QuantConfig Panel to support version 3.0

+++ prolinestudio-mzscope/src/main/java/fr/proline/mzscope/mzdb/MzdbRawFile.java
@@ -564,8 +564,8 @@ public class MzdbRawFile implements IRawFile {
         try {
             fr.profi.mzdb.model.Spectrum rawSpectrum = reader.getSpectrum((long) spectrumIndex);
             SpectrumData data = rawSpectrum.getData();
-            Map<Integer, DataEncoding> map = reader.getDataEncodingReader().getDataEncodingById();
-            DataEncoding encoding = reader.getSpectrumDataEncoding((long) spectrumIndex);
+            Map<Long, DataEncoding> map = reader.getDataEncodingReader().getDataEncodingById();
+            DataEncoding encoding = reader.getSpectrumDataEncoding(spectrumIndex);
             // TODO use mzdb-access ion_mobility branch to enable this feature
             // Spectrum.ScanType scanType = (encoding.getMode().equals(DataMode.CENTROID) || encoding.getMode().equals(DataMode.CENTROID_3D)) ?  Spectrum.ScanType.CENTROID : Spectrum.ScanType.PROFILE;
